+++ examples/config/channel-client.ts
@@ -1,8 +1,18 @@
-import { parseContractAddress, parseHexKey, parseOptional } from '../../sdk/src/env.js'
+import {
+  parseContractAddress,
+  parseHexKey,
+  parseNetworkId,
+  parseOptional,
+} from '../../sdk/src/env.js'
+import type { NetworkId } from '../../sdk/src/constants.js'
 
 export class Env {
   static get commitmentSecret(): string {
     return parseHexKey('COMMITMENT_SECRET')
+  }
+
+  static get network(): NetworkId {
+    return parseNetworkId()
   }
 
   static get serverUrl(): string {